First Commit
This code was written for the 2014 worlds robot. Our plans for the skills robot is very similar program wise, so I'm uploading this here because it is very similar to how the code will look.

I tried setting "liftAvoidTip" to a bool, but it didn't work when it was a bool. Not sure why... So the int will have to do.